refactor: separate user inventory and recipe search functionality into multiple components
