--- src/feats/InventoryProfile.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const InventoryProfile: React.FC = () => (
-  <div>
-    <h1>InventoryProfile View Working</h1>
-  </div>
-);
-
-export default InventoryProfile;
