+++ src/feats/UserInventoryAndRecipeSearch.tsx
@@ -2,33 +2,8 @@ import React, { useState, useEffect } from 'react';
 // import service from '../utils/api-service'; // Bring this back soon
 import { ISpoonacularRecipe } from '../utils/types';
 
-const RecipeSearch: React.FC = () => {
-  const [newFormEntry, setNewFormEntry] = useState<{ [key: string]: string }>({});
-  const handleFormChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    e.persist();
-    setNewFormEntry((prevState) => ({
-      ...prevState,
-      [e.target.name]: e.target.value,
-    }));
-  };
-
+const UserInventoryAndRecipeSearch: React.FC = () => {
   const [recipeSearchResults, setRecipeSearchResults] = useState<ISpoonacularRecipe[]>([]);
-
-  const handleFormSubmit = (e: React.MouseEvent<HTMLButtonElement>) => {
-    e.preventDefault();
-    fetch('/recipes-by-ingredients', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(newFormEntry),
-    })
-      .then((r: any) => r.json())
-      .then((r: any) => {
-        setRecipeSearchResults(r);
-      })
-      .catch((err: any) => console.log({ Error: err, Status: err.status, Message: err.message }));
-  };
 
   const [userAvailableInventory, setUserAvailableInventory] = useState<any[]>([]);
   const userId = 1; // Replace this w/ authenticated user's ID
@@ -97,21 +72,6 @@ const RecipeSearch: React.FC = () => {
             <div className="col col--12">
               <h3 className="card--title">Search Recipes by Ingredients</h3>
               <div className="card__body">
-                <input
-                  className="form-input margin-bottom--md"
-                  value={newFormEntry.listOfIngredients || ''}
-                  type="text"
-                  name="listOfIngredients"
-                  placeholder="List ingredients with commas"
-                  onChange={handleFormChange}
-                />
-                <button
-                  type="button"
-                  className="button button--outline button--primary"
-                  onClick={handleFormSubmit}
-                >
-                  Submit Search
-                </button>
                 <button
                   type="button"
                   className="button button--outline button--primary"
@@ -120,6 +80,7 @@ const RecipeSearch: React.FC = () => {
                   Submit Inventory Search
                 </button>
               </div>
+              <div>DISPLAY CURRENT AVAILABLE INVENTORY OF USER HERE</div>
             </div>
           </div>
         </div>
@@ -133,4 +94,4 @@ const RecipeSearch: React.FC = () => {
   );
 };
 
-export default RecipeSearch;
+export default UserInventoryAndRecipeSearch;
